still nailing down that test
git-svn-id: http://svn.apostrophenow.org/plugins/apostrophePlugin/trunk@72 c89fab08-2f9b-42bf-8176-bd3ad6bee547

+++ lib/task/apostropheMigratefrompkcontextcmsTask.class.php
@@ -324,7 +324,9 @@ BACK UP YOUR PROJECT BEFORE YOU RUN THIS SCRIPT, INCLUDING YOUR DATABASE.
       echo("Already exists\n");
       return;
     }
-    if ((is_dir($from) && file_exists($from . '/.svn')) || file_exists(dirname($from) . '/.svn'))
+    echo("Is it a directory: ");
+    echo(is_dir($from) . "\n");
+    if ((is_dir($from) && file_exists($from . '/.svn')) || ((!is_dir($from)) && file_exists(dirname($from) . '/.svn')))
     {
       $cmd = 'svn mv --parents ' . escapeshellarg($from) . ' ' . escapeshellarg($to);
       echo("$cmd\n");
